refactor(fs): consolidate containers & components
Move all presentational components to the top level components directory
and all container components to the top level containers directory.
