+++ app/components/Activity/Payment/Payment.js
