+++ app/components/Activity/Payment/index.js
