+++ app/components/Activity/index.js
@@ -1,0 +1,3 @@
+import Activity from './Activity'
+
+export default Activity
