+++ app/containers/App.js
@@ -78,7 +78,7 @@ import {
   activitySelectors
 } from 'reducers/activity'
 
-import App from '../components/App'
+import App from 'components/App'
 
 const mapDispatchToProps = {
   fetchTicker,
