--- app/routes/activity/index.js
@@ -1,3 +1,0 @@
-import ActivityContainer from './containers/ActivityContainer'
-
-export default ActivityContainer
