--- app/routes/app/components/App.scss
@@ -1,16 +1,0 @@
-@import '../../../styles/variables.scss';
-
-.content {
-	position: relative;
-	width: 70%;
-	display: inline-block;
-	vertical-align: top;
-	overflow-y: auto;
-}
-
-.titleBar {
-  background: $spacegrey;
-  height: 20px;
-  -webkit-user-select: none;
-  -webkit-app-region: drag;
-}
