--- app/routes/app/index.js
@@ -1,3 +1,0 @@
-import AppContainer from './containers/AppContainer'
-
-export default AppContainer
